+++ backend/src/models/dto/message/res/get-messages-list-res.dto.ts
@@ -1,0 +1,11 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+import { MessagesByChatResDto } from './messages-by-chat-res.dto';
+
+export class GetMessagesListResDto {
+	@ApiProperty({
+		nullable: false,
+		type: [MessagesByChatResDto],
+	})
+	list!: MessagesByChatResDto[];
+}
