+++ backend/src/models/repositories/chats.repository.ts
@@ -14,48 +14,6 @@ export class ChatsRepository extends Repository<ChatEntity> {
 		super(repository.target, repository.manager, repository.queryRunner);
 	}
 
-	async getMessagesByChat(chatId: string, memberId: string) {
-		return this.repository
-			.findOneOrFail({
-				select: {
-					id: true,
-					createdAt: true,
-					updatedAt: true,
-					messages: {
-						id: true,
-						createdAt: true,
-						updatedAt: true,
-						message: true,
-						member: {
-							id: true,
-							username: true,
-							profileImage: true,
-						},
-					},
-				},
-				where: {
-					id: chatId,
-				},
-				relations: {
-					messages: {
-						member: true,
-					},
-				},
-			})
-			.then((data) => {
-				const newMessages = data.messages.map((value) => {
-					return {
-						...value,
-						isMine: memberId === value.member.id ? true : false,
-					};
-				});
-				return {
-					...data,
-					messages: newMessages,
-				};
-			});
-	}
-
 	async createChat(): Promise<{ id: string }> {
 		const chat = await this.repository.insert({
 			id: uuidv4(),
